Rename helper/query_parser to helper/text_parser
