+++ helper/text_parser.js
