+++ test/unit/query/autocomplete.js
@@ -1,6 +1,6 @@
 
 var generate = require('../../../query/autocomplete');
-var parser = require('../../../helper/query_parser');
+var parser = require('../../../helper/text_parser');
 
 module.exports.tests = {};
